+++ src/model/tournament/tournament.model.ts
@@ -9,12 +9,15 @@
 // }
 
 export interface TournamentDto {
-	id: string;
+	id: string;				// created by the server
 	name: string;			// must be unique
 	city?: string;
 	country?: string;
 	month?: number;
 	year?: number;
 	rounds: number;
+	maxPlayers: number;
+	type: string; //roundRobin, swiss, elimination, match
 	players?: Array<number> // TournamentPlayerDto.id
+	state?: string; // planned* / scheduled / closed / underway / complete
 }