+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
+## [0.0.6] - underway
+### Feature
+* Add state to user: active / inactive 
+* Add start to tournament: planned / scheduled / underway / complete
+* Improve testing
+
 ## [0.0.5] - 2021 01 30
 ### Feature
 * Added the tournament entity